completed logout method with django's default logout
