+++ dww_backend/api/urls.py
@@ -6,5 +6,6 @@ urlpatterns = [
     path('test/', views.test, name='test'), 
     path('register-provider', views.register_provider, name='register provider'), 
     path('register/', views.register, name='register user'),
-    path('add-relationship/', views.add_relationship, name='add_relationship')
+    path('add-relationship/', views.add_relationship, name='add_relationship'),
+    path('logout/', views.logout_view, name='logout')
 ]
